UPSTREAM: media: cec: move cec_register_cec_notifier to cec-notifier.h
The cec_register_cec_notifier function was in media/cec.h, but it
has to be in cec-notifier.h.

While we are at it, also document it and add a stub function for when
the notifier is disabled or the CEC core code is unreachable.

Based on an earlier patch from Jose Abreu <Jose.Abreu@synopsys.com>.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@s-opensource.com>
(cherry picked from commit 5150418593015d280dab321da1fe3cea4ed3b693)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: I817186c710ca790ff3fd8315a11298f539a14b8c
Reviewed-on: https://chromium-review.googlesource.com/1021348
Commit-Ready: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Sean Paul <seanpaul@google.com>