CHROMIUM: HACK: drm/panel: simple: Fake compromise timings for nefario
This imaginary panel was conceived of to provide timings that will
work no matter which of the 5 possible panels gets stuffed on the
-rev0 nefarios that will be built.

This patch is intended to be reverted once crbug.com/750354 is fixed.
This patch should not be sent upstream.

---

These timings were generated with:

freq = 75000000
ideal_hz = 59.0

best_delta = 99999
min_v = max([771,  780,  793,  776, 778])
max_v = min([1023,  840, 1000, 830])
min_h = max([1420, 1590, 1456, 1426, 1520])
max_h = min([2047, 1692, 2000, 1620])

for vperiod in xrange(min_v, max_v + 1):
  for hperiod in xrange(min_h, max_h + 1):
    hz = float(freq) / (vperiod * hperiod)
    delta = abs(hz - ideal_hz)
    if delta < best_delta:
      best_delta = delta
      best_vperiod = vperiod
      best_hperiod = hperiod
      best_hz = hz

print best_hz, best_vperiod, best_hperiod

---

That gives us:

59.0003453487 793 1603

---

Note that the 75 MHz was picked because it is one of the few clock
rates that can be achieved via an integral divide of 600 MHz (GPLL) or
800 MHz (CPLL) due to the fact that we want to save VPLL for the
external display.  Basically the only reasonable pixel clocks we can
use for this sized panel are:

800 / 10. = 80 MHz
800 / 11. = 72.727 MHz
600 / 8.  = 75.0 MHz
600 / 9.  = 66.67 MHz

Picking 75.0 is the fastest pixel clock that all panels can handle.

We picked to aim for 59 Hz exactly since the refresh rate needs to be
an integral number of Hz and we couldn't make 60 Hz w/ all of our
compromises.

---

Note that the above timings only give us about 342 us of vblank timing
(if I'm doing my math right).  Currently DMC_MIN_VBLANK_NS is set to
300 us, so we're fairly close to the margin there.  Hopefully that
should be OK, but we should keep our eyes out for any possible
flickering.

---

BUG=chromium:750354, b:63537905
TEST=Compile

Change-Id: I50ce712220a4d3bab1a544f44254c2b43aae0145
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/594428
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>