CHROMIUM: arm64: dts: rockchip: update wifi power sequence for Scarlet
We've reduced the number of power rails on Scarlet, so there's a little
less flexibility here. Also, the module specification notes a little
different timings. I see:

TPVGL [1] = minimum 10ms

So we ensure 10ms between PD# goes high for 10ms before we init the PCIe
link.

I don't see a spec for how long we need to hold PD# low to ensure a
proper reset, so we just assert it "ASAP" in the pinctrl hogs, and then
assert it when we're ready to init the PCIe root complex.

This is similar (but not quite the same) as the base Gru reference
board, since we don't yet enforce a minimum PD# active time.

[1] power valid to PERST# input inactive

BUG=b:64116574, b:64444991
TEST=wifi can work on scarlet board

Change-Id: I32733bb7b04cdf0d6d49a4fe5fe5d7d6c7f84679
Signed-off-by: Lin Huang <hl@rock-chips.com>
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/593409
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>