HACK: hw_random: tpm-rng: don't access tpm when suspended
hwrng_fillfn thread in core hwrng driver may still be
generating initial randomness when the device is suspended.
In this case tpm_rng_read() may send a GetRandom message
to the tpm after the tpm state was saved in preparation to
suspend, and thus cause the subsequent resume to fail.

To prevent that, create a platform device for tpm-rng after
the tpm is available in the system, handle suspend-resume
callbacks and block tpm accesses while the device is
suspended.

A better method is implementing the tpm-rng 'plugin' for
hwrng driver as a part of the tpm driver itself - calling
hwrng_register() and handling suspend-resume there.
Thus this patch is considered temporary and marked as HACK.

BUG=b:63709621
BUG=b:63164064
BUG=b:62848911
TEST=Go through suspend-resume, check that tpm-rng is
     suspended before tpm0 and resumed after.

Change-Id: I969013aea277b2cf330a003a65cbba32eacaf623
Signed-off-by: Andrey Pronin <apronin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/596387
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Chung-yih Wang <cywang@chromium.org>
(cherry picked from commit 2feb20f413f7413c46e96e2bd13dc60137e3f680)
Reviewed-on: https://chromium-review.googlesource.com/607731