HACK: hw_random: tpm-rng: don't access tpm when suspended
hwrng_fillfn thread in core hwrng driver may still be
generating initial randomness when the device is suspended.
In this case tpm_rng_read() may send a GetRandom message
to the tpm after the tpm state was saved in preparation to
suspend, and thus cause the subsequent resume to fail.

To prevent that, create a platform device for tpm-rng after
the tpm is available in the system, handle suspend-resume
callbacks and block tpm accesses while the device is
suspended.

A better method is implementing the tpm-rng 'plugin' for
hwrng driver as a part of the tpm driver itself - calling
hwrng_register() and handling suspend-resume there.
Thus this patch is considered temporary and marked as HACK.

BUG=b:63709621
TEST=Go through suspend-resume, check that tpm-rng is
     suspended before tpm0 and resumed after.

Change-Id: I17e576482c84bc039d0e350b04b07577c4645bd5
Signed-off-by: Andrey Pronin <apronin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/611403
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
(cherry picked from commit 44b6ff0621af1f7b9632ca3ea2a0d1c4e219821f)
Reviewed-on: https://chromium-review.googlesource.com/614782