CHROMIUM: lib: lzo: clean-up by introducing COPY16
Most compilers should be able to merge adjacent loads/stores of sizes which
are less than but effect a multiple of a machine word size (in effect a
memcpy() of a constant amount). However the semantics of the macro are that
it just does the copy, the pointer increment is in the code, hence we see

    *a = *b
    a += 8
    b += 8
    *a = *b
    a += 8
    b += 8

This introduces a dependency between the two groups of statements which
seems to defeat said compiler optimizers and generate some very strange
sequences of addition and subtraction of address offsets (i.e. it is
overcomplicated).

Since COPY8 is only ever used to copy amounts of 16 bytes (in pairs), just
define COPY16 as COPY8,COPY8. We leave the definition to preserve the need
to do unaligned accesses to machine-sized words per the original code
intent, we just don't use it in the code proper.

COPY16 then gives us code like:

    *a = *b
    *(a+8) = *(b+8)
    a += 16
    b += 16

This seems to allow compilers to generate much better code by using
base register writeback or simply positively incrementing offsets which
seems to positively affect performance. It is, at least, fewer instructions
to do the same job.

BUG=chromium:907138
TEST=python -c 'l = [0] * 1024*1024*400; l2 = [0] * 1024*1024*400'
TEST=zramctl; dmesg
TEST=zramctl should show evidence of swapping using the lzo algorithm,
TEST=and dmesg should not report any errors

Change-Id: Ia6f2e90977a7f232139fba1b0555673b24e9a837
Signed-off-by: Matt Sealey <matt.sealey@arm.com>
Reviewed-on: https://chromium-review.googlesource.com/1307374
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>