CHROMIUM: lib: lzo: enable 64-bit CTZ on Arm
ARMv6 Thumb state introduced an RBIT instruction which, combined with CLZ
as present in ARMv5, introduces an extremely fast path for counting
trailing zeroes.

Enable the use of the GCC builtin for this on ARMv6+ with
CONFIG_THUMB2_KERNEL to ensure we get the 'new' instruction usage.

We do not bother enabling LZO_USE_CTZ64 support for ARMv5 as the builtin
code path does the same thing as the LZO_USE_CTZ32 code, only with more
register pressure.

BUG=chromium:907138
TEST=python -c 'l = [0] * 1024*1024*400; l2 = [0] * 1024*1024*400'
TEST=zramctl; dmesg
TEST=zramctl should show evidence of swapping using the lzo algorithm,
TEST=and dmesg should not report any errors

Change-Id: I3c0061fcfaa5f9e35e58f4df69fe304f46e1ddd7
Signed-off-by: Matt Sealey <matt.sealey@arm.com>
Reviewed-on: https://chromium-review.googlesource.com/1307375
Commit-Ready: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>