UPSTREAM: ipv6: fib: Unlink replaced routes from their nodes
When a route is deleted its node pointer is set to NULL to indicate it's
no longer linked to its node. Do the same for routes that are replaced.

This will later allow us to test if a route is still in the FIB by
checking its node pointer instead of its reference count.

BUG=b:64624479
TEST=Build and run

Change-Id: I90f6291da58b48178d92830fe86a5ac37d9e1843
Signed-off-by: Ido Schimmel <idosch@mellanox.com>
Signed-off-by: Jiri Pirko <jiri@mellanox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Guenter Roeck <groeck@chromium.org>
(cherry picked from commit 7483cea79957)
Reviewed-on: https://chromium-review.googlesource.com/749054
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>