drm/i915/ICL: Add pre_pll_enable hook for ICL and set DFLEXDPMLE in this hook
In case of Legacy DP connector on TypeC port (C, D, E or F), the
flex IO DPMLE register is set to maximum number of lanes since
there is no muxing with other controllers in this case.
This needs to be programmed before enabling the shared PLLs hence
add a pre_pll_enable hook for ICL and add this programming in that hook.

v6:
* s/ICL_PORT_TX_DFLEXDPMLE1/PORT_TX_DFLEXDPMLE1/
v5:
* Fix encoder type check (DK).
v4:
* Use < 8 instead of <= 7 in the for loop (Jani N)
v3:
* Add FIXME for some comments (Jani Nikula)
* Change the function name that sets the DPMLE register (Jani, Animesh)
v2:
* Add pre pll enable hook and move dflexdpmle programming
to that hook (Animesh)
* Add a check for encoder and port type to ensure
that we dont set it to max for type c and TBT (Animesh)

Cc: Jani Nikula <jani.nikula@linux.intel.com>
Cc: Animesh Manna <animesh.manna@intel.com>
Cc: Anusha Srivatsa <anusha.srivatsa@intel.com>
Cc: Dhinakaran Pandiyan <dhinakaran.pandiyan@intel.com>
Cc: Paulo Zanoni <paulo.r.zanoni@intel.com>
Signed-off-by: Manasi Navare <manasi.d.navare@intel.com>